Rename TokenStatusPage components
